# src/main/resources/swagger/docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cats-jooq",
    "version": "local",
    "description": "A sample application for demonstrating the use of the jooq library with the cats effect."
  },
  "tags": [
    {
      "name": "Users"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/{email}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "schema": {
              "type": [
                "string"
              ],
              "format": "email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "schema": {
              "type": [
                "string"
              ]
            },
            "description": "abc@abc.abc"
          }
        ],
        "responses": {
          "204": {},
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "schema": {
              "type": [
                "string"
              ],
              "format": "email"
            }
          },
          {
            "in": "query",
            "name": "lastName",
            "schema": {
              "type": [
                "string"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "no content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/stream": {
      "get": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Gender": {
        "type": [
          "string"
        ],
        "enum": [
          "male",
          "female"
        ]
      },
      "User": {
        "type": [
          "object"
        ],
        "properties": {
          "firstName": {
            "type": [
              "string"
            ]
          },
          "lastName": {
            "type": [
              "string"
            ]
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "dob": {
            "type": [
              "string"
            ],
            "format": "date"
          },
          "email": {
            "type": [
              "string"
            ],
            "format": "email"
          },
          "mobile": {
            "type": [
              "string"
            ],
            "format": "Phone Number"
          }
        },
        "required": [
          "firstName",
          "gender",
          "dob",
          "email"
        ]
      },
      "Error": {
        "type": [
          "object"
        ],
        "properties": {
          "isError": {
            "type": [
              "boolean"
            ]
          },
          "errorCode": {
            "type": [
              "string"
            ]
          }
        }
      }
    },
    "responses": {
      "400": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "500": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}